requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==2.4.1
attrs==25.3.0
botcity-maestro-sdk==0.9.0
cachetools==6.2.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.7
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.26.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jiter==0.11.0
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.2.1
matplotlib-inline==0.1.7
more-itertools==10.3.0
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.0.1
openai==2.0.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patrimar_dependencies @ git+https://github.com/RenanMGX/dependencies_patrimar.git@24a88786f53cfe7655666cebe58b8253a6943ae8
pillow==11.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.32.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
striprtf==0.0.26
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xlwings==0.33.1
zipp==3.23.0
